\name{RainFallExample}
\alias{RainFallExample}
\title{The Rainfall Example Data Set-Runoff Erosivity Factor}
\description{
  Data set for determination of rainfall runoff erosivity factor
}
\details{
  Data should be organized into notebooks or spreadsheets, where the first column
  being the date (dd/mm/yyyy), the second one, the hour column (hh:mm) and the third column the precipitation in mm. The
 dot (.) must be used as decimal separator.
 
 The results of package is: Monthy value (Final rainfall sum, Total erosivity [EI30], Number erosive, Number non erosive, Total erosive sum, Total non erosive sum).Annual value (Final rainfall sum, Total erosivity [EI30], Number erosive, Number non erosive, Total erosive sum, Total non erosive sum). R factor (average total erosivity)
}
\keyword{
    Erosivity index; R-Factor; Universal Soil Loss Equation-USLE
}
\references{
    Wischmeier, W. & Smith, D. (1978). Predicting rainfall erosion losses: a guide to conservation planning. Agricultural Handbook No. 537. U.S. Department of Agriculture,Washington DC, USA.
}

